airflow/requirements.txt
kafka-python
pandas

scikit-learn
joblib
google-cloud-bigquery
pyarrow>=10.0.1,<10.1.0
